+++ pharmacy.test.js
@@ -1,4 +1,5 @@
-import { Drug, Pharmacy } from './pharmacy'
+import { Pharmacy } from './pharmacy'
+import { Drug } from './drug'
 
 describe('Pharmacy', () => {
   function expectUpdatedPharmacy(updatedBenefitValue, expectedTestDrug) {
